requirements.txt
# chromadb needs python <=3.12 
# python3.12 -m venv .venv 
# pip install -r requirements.txt
pydantic
ollama
pypdf
chromadb
python-docx
python-pptx